更新navbar,分類頁碼完成,更新CSS

+++ product_list.php
@@ -1,115 +1,168 @@
 <?php
-include __DIR__.'/init.php';
+include __DIR__ . '/init.php';
 
 // ------------------------頁數設定-----------------------
-// $where = 'WHERE 1 ';
 
 $perPage = 5;
 $page = isset($_GET['page']) ? intval($_GET['page']) : 1;
-
-$totalRows = $pdo->query("SELECT count(1) FROM `products_food` WHERE 1;")
-    ->fetch(PDO::FETCH_NUM)[0];
-
-$totalPages = ceil($totalRows/$perPage);
-
-$sql = sprintf("SELECT * FROM `products_food` ORDER BY sid DESC LIMIT %s, %s", ($page-1)*$perPage, $perPage);
-
-$qs = [];
-// ------------------------分類資料呈現--------------------
 $cate = isset($_GET['cate']) ? intval($_GET['cate']) : 0;
+$qs = [];
 
-if(!$cate){
-    $sql = sprintf("SELECT * FROM `products_food` ORDER BY sid DESC LIMIT %s, %s", ($page-1)*$perPage, $perPage);
-}else{
+$where = 'WHERE 1 ';
 
+// 如果沒有類別,sql就設定全部,如果有sql就設定照類別分
+if (!$cate) {
+    $sql = sprintf("SELECT * FROM `products_food` ORDER BY sid DESC LIMIT %s, %s", ($page - 1) * $perPage, $perPage);
+    $stmt = $pdo->query($sql);
+    $rows = $stmt->fetchAll();
+} else {
+    $where .= "AND cate_sid = $cate";
+    $qs['cate'] = $cate;
     $sql = "SELECT * FROM `products_food` WHERE `cate_sid`=$cate";
-
+    $rows = $pdo->query($sql)->fetchALL();
 }
 
-$row = $pdo->query($sql)->fetchALL();
+
+// 所有資料
+$totalSql = "SELECT count(1) FROM `products_food` $where;";
+// 總筆數
+$totalRows = $pdo->query("SELECT count(1) FROM `products_food` $where;")
+    ->fetch(PDO::FETCH_NUM)[0];
+
+$totalPages = ceil($totalRows / $perPage);
+
 
 ?>
 
-<?php include __DIR__.'./partials/html-head.php' ?>
-<?php include __DIR__.'./partials/nav_bar.php' ?>
+<?php include __DIR__ . './partials/html-head.php' ?>
+<?php include __DIR__ . './partials/nav_bar.php' ?>
 
 <style>
-    .list{
+    .list {
         margin-top: 0px;
     }
 
-    .intr{
+    .intr {
         width: 350px;
     }
-    .content{
+
+    .content {
         width: 150px;
     }
-    .name{
+
+    .name {
         width: 150px;
     }
-    .sid{
-        width: 50px;
-    }
-    .brand{
+
+    .brand {
         width: 90px;
     }
 
-    .page-btn{
+    .page-btn {
         margin-top: 10px;
     }
 
-    img{
+    img {
         width: 80px;
     }
-    i{
+
+    i {
         font-size: 20px;
     }
-    .icon{
+
+    .icon {
         text-align: center;
     }
-    .fa-edit{
+
+    .fa-edit {
         color: darkblue;
     }
-    .fa-trash-alt{
+
+    .fa-trash-alt {
         color: darkred;
+    }
+
+
+    .btn-row {
+        justify-content: space-between;
+        display: flex;
+        align-items: center;
+    }
+
+    .btnwrap {
+        display: flex;
+        flex-direction: row;
+        /* justify-content: ; */
+        width: 40%;
+    }
+
+    .btnwrap div {
+        width: 100px;
+        color: darkblue;
+    }
+
+    .page {
+        width: 30%;
+        display: flex;
+        flex-direction: row-reverse;
+    }
+
+    .create a {
+        /* text-decoration: none; */
+        color: darkred;
+        font-weight: bold;
     }
 </style>
 
-<!-- 頁碼 -->
-<div class="container">
-    <div class="row">
-        <div class="col">
-            <nav aria-label="Page navigation example">
-                <ul class="pagination page-btn">
-                    
-                    <!-- 前一頁icon -->
-                    <li class="page-item <?= $page<=1? 'disabled':'' ?>">
-                        <a class="page-link"
-                        href="?<?php $qs['page']=$page-1; echo http_build_query($qs); ?>">
-                            <i class="fas fa-angle-left"></i>
+
+<div class="container" d-flex>
+    <div class="btn-row">
+        <!-- 分類按鈕 -->
+        <div class="btnwrap ">
+            <div class="create">
+                <a href="product_create.php">新增商品</a>
+            </div>
+            <div class="">
+                <a href="?">所有商品</a>
+            </div>
+            <div class="">
+                <a href="?cate=1">快速上桌</a>
+            </div>
+            <div class="">
+                <a href="?cate=2">健身專區</a>
+            </div>
+            <div class="">
+                <a href="?cate=3">嚴選食材</a>
+            </div>
+        </div>
+        <!-- 頁碼 -->
+        <div class="page">
+            <ul class="pagination page-btn">
+                <!-- 前一頁icon -->
+                <li class="page-item <?= $page <= 1 ? 'disabled' : '' ?>">
+                    <a class="page-link" href="?<?php $qs['page'] = $page - 1;
+                                                echo http_build_query($qs); ?>">
+                        <i class="fas fa-angle-left"></i>
+                    </a>
+                </li>
+                <!-- --------頁數-------- -->
+                <?php for ($i = 1; $i <= $totalPages; $i++) :
+                    $qs['page'] = $i;
+                ?>
+                    <li class="page-item <?= $i == $page ? 'active' : '' ?>">
+                        <a class="page-link" href="?<?= http_build_query($qs) ?>">
+                            <?= $i ?>
                         </a>
                     </li>
-                    <!-- --------頁數-------- -->
-                    <?php for($i=$page-5; $i<=$page+5; $i++):
-                            if($i>=1 and $i<=$totalPages):
-                                $qs['page'] = $i;
-                        ?>
-                    <li class="page-item <?= $i==$page ? 'active' : '' ?>">
-                        <a class="page-link" 
-                        href="?<?= http_build_query($qs)?>">
-                        <?= $i ?>
-                        </a>
-                    </li>
-                    <?php endif;
-                        endfor; ?>
-    
-                    <li class="page-item <?= $page>=$totalPages ? 'disabled' : '' ?>">
-                        <a class="page-link" href="?<?php $qs['page']=$page+1; echo http_build_query($qs);?>">
-                            <i class="fas fa-angle-right"></i>
-                        </a>
-                    </li>
-                </ul>
-            </nav>
+                <?php endfor; ?>
+
+                <li class="page-item <?= $page >= $totalPages ? 'disabled' : '' ?>">
+                    <a class="page-link" href="?<?php $qs['page'] = $page + 1;
+                                                echo http_build_query($qs); ?>">
+                        <i class="fas fa-angle-right"></i>
+                    </a>
+                </li>
+            </ul>
         </div>
     </div>
 </div>
@@ -117,12 +170,12 @@ $row = $pdo->query($sql)->fetchALL();
 <!-- 表單資訊 -->
 <div class="container list">
     <div class="row">
-    <div class="col">
+        <div class="col">
             <table class="table table-striped table-bordered">
                 <thead>
                     <tr>
                         <th scope="col">刪除</th>
-                        <th class="sid" scope="col">sid</th>
+                        <!-- <th class="sid" scope="col">sid</th> -->
                         <th class="name" scope="col">商品名稱</th>
                         <th class="intr" scope="col">商品介紹</th>
                         <th class="brand" scope="col">品牌</th>
@@ -134,22 +187,21 @@ $row = $pdo->query($sql)->fetchALL();
                     </tr>
                 </thead>
                 <tbody>
-                    <?php foreach($row as $r): ?>
-                        <tr data-sid="<?= $r['sid'] ?>" >
+                    <?php foreach ($rows as $r) : ?>
+                        <tr data-sid="<?= $r['sid'] ?>">
                             <td class="icon">
-                                <a href="product_delete.php?sid=<?= $r['sid'] ?>"
-                                onclick="return confirm('確定要刪除<?= $r['name'] ?>的商品資料嗎')">
+                                <a href="product_delete.php?sid=<?= $r['sid'] ?>" onclick="return confirm('確定要刪除<?= $r['name'] ?>的商品資料嗎')">
                                     <i class="fas fa-trash-alt"></i>
                                 </a>
                             </td>
-                            <td><?= $r['sid'] ?></td>
+                            <!-- <td><?= $r['sid'] ?></td> -->
                             <td><?= $r['name'] ?></td>
                             <td><?= $r['introduction'] ?></td>
                             <td><?= $r['brand'] ?></td>
                             <td><?= $r['price'] ?></td>
                             <td><?= $r['content'] ?></td>
                             <td><img src="./img/<?= $r['img'] ?>" alt=""></td>
-                            <td class="icon" >
+                            <td class="icon">
                                 <a href="product_edit.php?sid=<?= $r['sid'] ?>">
                                     <i class="fas fa-edit"></i>
                                 </a>
@@ -161,13 +213,5 @@ $row = $pdo->query($sql)->fetchALL();
         </div>
     </div>
 </div>
-<?php include __DIR__.'./partials/scripts.php' ?>
-<?php include __DIR__.'./partials/html-foot.php' ?>
-
-
-
-
-
-
-
-
+<?php include __DIR__ . './partials/scripts.php' ?>
+<?php include __DIR__ . './partials/html-foot.php' ?>
